Passed nestest with success

+++ INST/INCDEC.h
@@ -6,15 +6,15 @@
 u8 DEC(u8 mem);
 void DEC_ZeroPage(u8 addr);
 void DEC_ZeroPageX(u8 addr);
-void DEC_Absolute(u8 addr);
-void DEC_AbsoluteX(u8 addr);
+void DEC_Absolute(u16 addr);
+void DEC_AbsoluteX(u16 addr);
 void DEX();
 void DEY();
 u8 INC(u8 mem);
 void INC_ZeroPage(u8 addr);
 void INC_ZeroPageX(u8 addr);
-void INC_Absolute(u8 addr);
-void INC_AbsoluteX(u8 addr);
+void INC_Absolute(u16 addr);
+void INC_AbsoluteX(u16 addr);
 void INX();
 void INY();
